sdks/multion-py/pyproject.toml
[tool.poetry]
name = "multion"
version = "0.3.7"
description = "MULTION API"
authors = ["Div Garg"]
include = ["multion/*"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
requests-oauthlib = "^1.3.1"
requests = "^2.31.0"
openai = "^1.3.0"
cryptography = "^41.0.1"
ipython = "^8.1.0"
cognitojwt = "^1.4.1"
pillow = "^10.0.1"
pytesseract = "^0.3.10"
deprecated = "^1.2.14"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
